requirements.txt
boto3
botocore
awscli
bitmath==1.3.1.2
cachetools==2.0.1
cement==2.10.2
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.7
dill==0.2.7.1
docutils==0.14
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gapic-google-cloud-pubsub-v1==0.15.4
gapic-google-cloud-spanner-admin-database-v1==0.15.3
gapic-google-cloud-spanner-admin-instance-v1==0.15.3
gapic-google-cloud-spanner-v1==0.15.3
google-api-python-client==1.6.4
google-auth==1.1.1
google-auth-httplib2==0.0.2
google-cloud==0.27.0
google-cloud-bigquery==0.26.0
google-cloud-bigtable==0.26.0
google-cloud-core==0.26.0
google-cloud-datastore==1.2.0
google-cloud-dns==0.26.0
google-cloud-error-reporting==0.26.0
google-cloud-language==0.27.0
google-cloud-logging==1.2.0
google-cloud-monitoring==0.26.0
google-cloud-pubsub==0.27.0
google-cloud-resource-manager==0.26.0
google-cloud-runtimeconfig==0.26.0
google-cloud-spanner==0.26.0
google-cloud-speech==0.28.0
google-cloud-storage==1.3.2
google-cloud-translate==1.1.0
google-cloud-videointelligence==0.25.0
google-cloud-vision==0.26.0
google-gax==0.15.15
google-resumable-media==0.2.3
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.6.3
httplib2==0.10.3
idna==2.6
ipaddr==2.2.0
Jinja2==2.9.6
jmespath==0.9.3
MarkupSafe==1.0
monotonic==1.3
oauth2client==4.1.2
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
proto-google-cloud-pubsub-v1==0.15.4
proto-google-cloud-spanner-admin-database-v1==0.15.3
proto-google-cloud-spanner-admin-instance-v1==0.15.3
proto-google-cloud-spanner-v1==0.15.3
protobuf==3.4.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.11
six==1.11.0
SQLAlchemy==1.1.14
tabulate==0.8.1
tenacity==4.4.0
transitions==0.6.1
uritemplate==3.0.0
urllib3==1.22
python-daemon==2.1.2
alabaster==0.7.10
argh==0.26.2
Babel==2.5.3
contexttimer==0.3.3
fasteners==0.14.1
imagesize==1.0.0
livereload==2.5.1
lockfile==0.12.2
packaging==17.1
pathtools==0.1.2
port-for==0.3.1
Pygments==2.2.0
pyparsing==2.2.0
python-daemon==2.1.2
pytz==2018.3
snowballstemmer==1.2.1
Sphinx==1.7.1
sphinx-autobuild==0.7.1
sphinxcontrib-websupport==1.0.1
sqlitedict==1.5.0
tornado==5.0
watchdog==0.8.3
